test basic char support
This patch adds a round trip test for the basic
char printing support adding in the last patch.
It does not test any sort of line handling.

We also add the test data files from libvterm.
Eventually, we'll want to write a parser for this
data and use it to test our virtual terminal engine,
but we probably won't get there for a while.

Also, I ran rustfmt for the first time.