add grid view module
This patch adds a new grid view module that will allow us to determine
which cells go where on the logical grid. We'll need this in order to
handle certain movement commands and whatnot.

The fragmented representation doesn't make a ton of sense now that
we are storing cell references rather than slices of cells, but it
was originally intended to be able to handle directly storing slices
into the underlying scrollback. It is probably worth thinking about
how to get back to being able to do that.

I'm also kinda worried that taking this approach of generating a grid
view of things from the underlying scrollback will make handling
control chars really slow when there is a large scrollback buffer.
Maybe we should somehow simultaniously maintain a grid and a scrollback,
with cells kinda belonging to both at once somehow.

I'm not super happy with this, but hopefully it is a step in roughly
the right direction.