implement new grid data model
This patch adds a new grid data model that should be able to
handle control codes more efficiently. The spiky list is going to
have problems, and this lays a better foundation. We should retain
pretty much the same memory usage patterns even though we'll be
using a grid as backing (because grid lines are short unless they
are fully used). This should shift the cost from control
code handling to reflow. Reflow happens much less frequently, so
this is a better tradeoff.

At some point I might want to try to optimize the memory further
with something like a bump allocator or something, but this is
probably fine for now.

I still need to update the main lib file to rip out its use of
scrollback.rs and switch to the new grid view.