add underline support
This patch adds underline support. In the process,
I completely overhauled the way attribute and control code handling
is done. The new approach is more data-oriented. It is more allocation
heavy and therefore probably slower, but the slowness will only
kick in when generating the content blob not as part of the main
process loop, so I don't think it matters too much. If we wind up
seeing noticable lag during session restore using this approach
we can circle back an optimize, but for now I'd like to use an
approach that is easier to implement and extend. I think this
approach may also be more amenable to reading codes directly
from terminfo rather than having them hardcoded, which I would
like to do eventually.