WriteBuf -> AsTermInput
This patch changes the WriteBuf trait to be called AsTermInput.
The old trait name is too generic and doesn't really explain
what the trait really does. Now, the name is more reflective
of the purpose of the trait.

I also deleted a bunch of the control code structs that I copied
from vt100 because they don't really make sense anymore given the
new data oriented design.